api/requirements.txt
fastapi 
uvicorn 
python-multipart 
requests
mysql-connector-python